+++ drivers/wifi/siwx91x/CMakeLists.txt
@@ -2,7 +2,7 @@
 # Copyright (c) 2024 Silicon Laboratories Inc.
 
 if(CONFIG_WIFI_SILABS_SIWX91X)
-  zephyr_library_sources(siwx91x_wifi.c)
+  zephyr_library_sources(siwx91x_wifi.c siwx91x_wifi_scan.c)
 
   if(CONFIG_WIFI_SILABS_SIWX91X_NET_STACK_OFFLOAD)
     zephyr_library_sources(siwx91x_wifi_socket.c)
